# Optimize notify of job listeners