--- a/src/Quartz.Tests.Unit/Core/QuartzSchedulerTest.cs
+++ b/src/Quartz.Tests.Unit/Core/QuartzSchedulerTest.cs
@@ -20,9 +20,9 @@
 #endregion
 
 using System;
+using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Globalization;
-using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -34,6 +34,7 @@ using Quartz.Core;
 using Quartz.Impl;
 using Quartz.Impl.Triggers;
 using Quartz.Job;
+using Quartz.Simpl;
 using Quartz.Spi;
 
 namespace Quartz.Tests.Unit.Core
@@ -142,6 +143,158 @@ namespace Quartz.Tests.Unit.Core
             // expect unschedule and schedule
             A.CallTo(() => listener.JobUnscheduled(new TriggerKey(TriggerName, TriggerGroup), A<CancellationToken>._)).MustHaveHappened();
             A.CallTo(() => listener.JobScheduled(jobTrigger, A<CancellationToken>._)).MustHaveHappened();
+        }
+
+        [Test]
+        [Ignore("Flaky in CI")]
+        public void CurrentlyExecutingJobs()
+        {
+            IReadOnlyCollection<IJobExecutionContext> executingJobs;
+
+            var scheduler = CreateQuartzScheduler("A", "B", 5);
+
+            executingJobs = scheduler.CurrentlyExecutingJobs;
+            Assert.AreEqual(0, executingJobs.Count);
+
+            scheduler.Start().GetAwaiter().GetResult();
+
+            executingJobs = scheduler.CurrentlyExecutingJobs;
+            Assert.AreEqual(0, executingJobs.Count);
+
+            ScheduleJobs<DelayedJob>(scheduler, 3, true, false, 1, TimeSpan.FromMilliseconds(1), 1);
+            ScheduleJobs<DelayedJob>(scheduler, 1, true, false, 1, TimeSpan.FromMilliseconds(1), 0);
+
+            Thread.Sleep(150);
+
+            executingJobs = scheduler.CurrentlyExecutingJobs;
+            Assert.AreEqual(4, executingJobs.Count);
+
+            Thread.Sleep(150);
+
+            executingJobs = scheduler.CurrentlyExecutingJobs;
+            Assert.AreEqual(3, executingJobs.Count);
+
+            Thread.Sleep(300);
+
+            executingJobs = scheduler.CurrentlyExecutingJobs;
+            Assert.AreEqual(0, executingJobs.Count);
+
+            scheduler.Shutdown(true).GetAwaiter().GetResult();
+        }
+
+        [Test]
+        [Ignore("Flaky in CI")]
+        public void NumJobsExecuted()
+        {
+            var scheduler = CreateQuartzScheduler("A", "B", 5);
+
+            Assert.AreEqual(0, scheduler.NumJobsExecuted);
+
+            scheduler.Start().GetAwaiter().GetResult();
+
+            Assert.AreEqual(0, scheduler.NumJobsExecuted);
+
+            ScheduleJobs<DelayedJob>(scheduler, 3, true, false, 1, TimeSpan.FromMilliseconds(1), 1);
+            ScheduleJobs<DelayedJob>(scheduler, 1, true, false, 1, TimeSpan.FromMilliseconds(1), 0);
+
+            Thread.Sleep(150);
+
+            Assert.AreEqual(4, scheduler.NumJobsExecuted);
+
+            Thread.Sleep(150);
+
+            Assert.AreEqual(7, scheduler.NumJobsExecuted);
+
+            Thread.Sleep(200);
+
+            Assert.AreEqual(7, scheduler.NumJobsExecuted);
+
+            scheduler.Shutdown(true).GetAwaiter().GetResult();
+        }
+
+        private static void ScheduleJobs<T>(QuartzScheduler scheduler,
+                                            int jobCount,
+                                            bool disableConcurrentExecution,
+                                            bool persistJobDataAfterExecution,
+                                            int triggersPerJob,
+                                            TimeSpan repeatInterval,
+                                            int repeatCount)
+        {
+            var triggersByJob = new Dictionary<IJobDetail, IReadOnlyCollection<ITrigger>>();
+
+            for (var i = 0; i < jobCount; i++)
+            {
+                var job = CreateJobDetail(typeof(QuartzSchedulerTest).Name,
+                                          typeof(T),
+                                          disableConcurrentExecution,
+                                          persistJobDataAfterExecution);
+
+                var triggers = new ITrigger[triggersPerJob];
+                for (var j = 0; j < triggersPerJob; j++)
+                {
+                    triggers[j] = CreateTrigger(job, repeatInterval, repeatCount);
+                }
+
+                triggersByJob.Add(job, triggers);
+            }
+
+            scheduler.ScheduleJobs(triggersByJob, false).GetAwaiter().GetResult();
+        }
+
+
+        private static QuartzScheduler CreateQuartzScheduler(string name, string instanceId, int threadCount)
+        {
+            var threadPool = new DefaultThreadPool { MaxConcurrency = threadCount };
+            threadPool.Initialize();
+
+            QuartzSchedulerResources res = new QuartzSchedulerResources
+            {
+                Name = name,
+                InstanceId = instanceId,
+                ThreadPool = threadPool,
+                JobRunShellFactory = new StdJobRunShellFactory(),
+                JobStore = new RAMJobStore(),
+                IdleWaitTime = TimeSpan.FromMilliseconds(10),
+                MaxBatchSize = threadCount,
+                BatchTimeWindow = TimeSpan.FromMilliseconds(10)
+            };
+
+            var scheduler = new QuartzScheduler(res);
+            scheduler.JobFactory = new SimpleJobFactory();
+            return scheduler;
+        }
+
+        private static ITrigger CreateTrigger(IJobDetail job, TimeSpan repeatInterval, int repeatCount)
+        {
+            return TriggerBuilder.Create()
+                                 .ForJob(job)
+                                 .WithSimpleSchedule(
+                                     sb => sb.WithRepeatCount(repeatCount)
+                                             .WithInterval(repeatInterval)
+                                             .WithMisfireHandlingInstruction(MisfireInstruction.IgnoreMisfirePolicy))
+                                 .Build();
+        }
+
+        private static IJobDetail CreateJobDetail(string group,
+                                                  Type jobType,
+                                                  bool disableConcurrentExecution,
+                                                  bool persistJobDataAfterExecution)
+        {
+            return JobBuilder.Create(jobType)
+                             .WithIdentity(Guid.NewGuid().ToString(), group)
+                             .DisallowConcurrentExecution(disableConcurrentExecution)
+                             .PersistJobDataAfterExecution(persistJobDataAfterExecution)
+                             .Build();
+        }
+
+        public class DelayedJob : IJob
+        {
+            private static TimeSpan _delay = TimeSpan.FromMilliseconds(200);
+
+            public async Task Execute(IJobExecutionContext context)
+            {
+                await Task.Delay(_delay).ConfigureAwait(false);
+            }
         }
     }
 }