--- a/src/Quartz/Core/ListenerManagerImpl.cs
+++ b/src/Quartz/Core/ListenerManagerImpl.cs
@@ -11,11 +11,13 @@ namespace Quartz.Core
     /// </summary>
     public class ListenerManagerImpl : IListenerManager
     {
-        private readonly OrderedDictionary<string, IJobListener> globalJobListeners = new OrderedDictionary<string, IJobListener>(10);
+        private readonly object globalJobListenerLock = new object();
+
+        private OrderedDictionary<string, IJobListener>? globalJobListeners;
 
         private readonly OrderedDictionary<string, ITriggerListener> globalTriggerListeners = new OrderedDictionary<string, ITriggerListener>(10);
 
-        private readonly Dictionary<string, List<IMatcher<JobKey>>> globalJobListenersMatchers = new Dictionary<string, List<IMatcher<JobKey>>>(10);
+        private Dictionary<string, List<IMatcher<JobKey>>>? globalJobListenersMatchers;
 
         private readonly Dictionary<string, List<IMatcher<TriggerKey>>> globalTriggerListenersMatchers = new Dictionary<string, List<IMatcher<TriggerKey>>>(10);
 
@@ -23,7 +25,9 @@ namespace Quartz.Core
 
         public void AddJobListener(IJobListener jobListener, params IMatcher<JobKey>[] matchers)
         {
-            AddJobListener(jobListener, new List<IMatcher<JobKey>>(matchers));
+            IReadOnlyCollection<IMatcher<JobKey>> matchersCollection = matchers;
+
+            AddJobListener(jobListener, matchersCollection);
         }
 
         public void AddJobListener(IJobListener jobListener, IReadOnlyCollection<IMatcher<JobKey>> matchers)
@@ -33,37 +37,56 @@ namespace Quartz.Core
                 throw new ArgumentException("JobListener name cannot be empty.");
             }
 
-            lock (globalJobListeners)
+            lock (globalJobListenerLock)
             {
+                globalJobListeners ??= new OrderedDictionary<string, IJobListener>();
                 globalJobListeners[jobListener.Name] = jobListener;
 
-                List<IMatcher<JobKey>> matchersL = new List<IMatcher<JobKey>>();
                 if (matchers != null && matchers.Count > 0)
                 {
-                    matchersL.AddRange(matchers);
+                    // Add or replace matchers for the job listener
+                    globalJobListenersMatchers ??= new Dictionary<string, List<IMatcher<JobKey>>>();
+                    globalJobListenersMatchers[jobListener.Name] = new List<IMatcher<JobKey>>(matchers);
                 }
                 else
                 {
-                    matchersL.Add(EverythingMatcher<JobKey>.AllJobs());
+                    // Remove any registered matchers for the job listener
+                    RemoveJobListenerMatchers(jobListener.Name);
                 }
-
-                globalJobListenersMatchers[jobListener.Name] = matchersL;
             }
         }
 
         public bool AddJobListenerMatcher(string listenerName, IMatcher<JobKey> matcher)
         {
-            if (matcher == null)
+            if (listenerName == null)
             {
-                throw new ArgumentException("Non-null value not acceptable.");
+                throw new ArgumentNullException(nameof(listenerName));
             }
 
-            lock (globalJobListeners)
+            if (matcher == null)
             {
-                if (!globalJobListenersMatchers.TryGetValue(listenerName, out var matchers))
+                throw new ArgumentNullException(nameof(matcher));
+            }
+
+            lock (globalJobListenerLock)
+            {
+                if (globalJobListenersMatchers == null || !globalJobListenersMatchers.TryGetValue(listenerName, out var matchers))
                 {
-                    return false;
+                    // Return false if no job listener is registered with the specified name
+                    if (globalJobListeners == null || !globalJobListeners.ContainsKey(listenerName))
+                    {
+                        return false;
+                    }
+
+                    // We may be adding the first matcher for any job listener, so make sure globalJobListenersMatchers
+                    // is initialized
+                    globalJobListenersMatchers ??= new Dictionary<string, List<IMatcher<JobKey>>>();
+
+                    // We're adding the first matcher for the specified job listener
+                    matchers = new List<IMatcher<JobKey>>();
+                    globalJobListenersMatchers.Add(listenerName, matchers);
                 }
+
                 matchers.Add(matcher);
                 return true;
             }
@@ -71,26 +94,54 @@ namespace Quartz.Core
 
         public bool RemoveJobListenerMatcher(string listenerName, IMatcher<JobKey> matcher)
         {
-            if (matcher == null)
+            if (listenerName == null)
             {
-                throw new ArgumentException("Non-null value not acceptable.");
+                throw new ArgumentNullException(nameof(listenerName));
             }
 
-            lock (globalJobListeners)
+            if (matcher == null)
             {
-                if (!globalJobListenersMatchers.TryGetValue(listenerName, out var matchers))
+                throw new ArgumentNullException(nameof(matcher));
+            }
+
+            if (globalJobListenersMatchers == null)
+            {
+                return false;
+            }
+
+            lock (globalJobListenerLock)
+            {
+                if (globalJobListenersMatchers == null || !globalJobListenersMatchers.TryGetValue(listenerName, out var matchers))
                 {
                     return false;
                 }
-                return matchers.Remove(matcher);
+
+                var removed = matchers.Remove(matcher);
+
+                if (removed && matchers.Count == 0)
+                {
+                    RemoveJobListenerMatchers(listenerName);
+                }
+
+                return removed;
             }
         }
 
         public IReadOnlyCollection<IMatcher<JobKey>>? GetJobListenerMatchers(string listenerName)
         {
-            lock (globalJobListeners)
+            if (listenerName == null)
             {
-                if (!globalJobListenersMatchers.TryGetValue(listenerName, out var matchers))
+                throw new ArgumentNullException(nameof(listenerName));
+            }
+
+            if (globalJobListenersMatchers == null)
+            {
+                return null;
+            }
+
+            lock (globalJobListenerLock)
+            {
+                if (globalJobListenersMatchers == null || !globalJobListenersMatchers.TryGetValue(listenerName, out var matchers))
                 {
                     return null;
                 }
@@ -101,46 +152,103 @@ namespace Quartz.Core
 
         public bool SetJobListenerMatchers(string listenerName, IReadOnlyCollection<IMatcher<JobKey>> matchers)
         {
-            if (matchers == null)
+            if (listenerName == null)
             {
-                throw new ArgumentException("Non-null value not acceptable.");
+                throw new ArgumentNullException(nameof(listenerName));
             }
 
-            lock (globalJobListeners)
+            if (matchers == null)
             {
-                if (!globalJobListenersMatchers.TryGetValue(listenerName, out _))
+                throw new ArgumentNullException(nameof(matchers));
+            }
+
+            lock (globalJobListenerLock)
+            {
+                if (globalJobListeners == null || !globalJobListeners.ContainsKey(listenerName))
                 {
                     return false;
                 }
 
-                globalJobListenersMatchers[listenerName] = new List<IMatcher<JobKey>>(matchers);
+                if (matchers.Count == 0)
+                {
+                    RemoveJobListenerMatchers(listenerName);
+                }
+                else
+                {
+                    globalJobListenersMatchers ??= new Dictionary<string, List<IMatcher<JobKey>>>();
+                    globalJobListenersMatchers[listenerName] = new List<IMatcher<JobKey>>(matchers);
+                }
+
                 return true;
             }
         }
 
         public bool RemoveJobListener(string name)
         {
-            lock (globalJobListeners)
+            if (name == null)
             {
-                return globalJobListeners.Remove(name);
+                throw new ArgumentNullException(nameof(name));
+            }
+
+            if (globalJobListeners == null)
+            {
+                return false;
+            }
+
+            lock (globalJobListenerLock)
+            {
+                if (globalJobListeners == null)
+                {
+                    return false;
+                }
+
+                var removed = globalJobListeners.Remove(name);
+
+                // When we've removed a job listener, make sure to also remove associated matchers
+                if (removed)
+                {
+                    RemoveJobListenerMatchers(name);
+
+                    if (globalJobListeners.Count == 0)
+                    {
+                        globalJobListeners = null;
+                    }
+                }
+
+                return removed;
             }
         }
 
-        public IReadOnlyCollection<IJobListener> GetJobListeners()
+        public IJobListener[] GetJobListeners()
         {
-            lock (globalJobListeners)
+            if (globalJobListeners == null)
             {
-                return globalJobListeners.Count > 0 
-                    ? new List<IJobListener>(globalJobListeners.Values) 
-                    : EmptyReadOnlyCollection<IJobListener>.Instance;
+                return Array.Empty<IJobListener>();
+            }
+
+            lock (globalJobListenerLock)
+            {
+                return globalJobListeners != null ? globalJobListeners.Values.ToArray()
+                                                  : Array.Empty<IJobListener>();
             }
         }
 
         public IJobListener GetJobListener(string name)
         {
-            lock (globalJobListeners)
+            if (name == null)
             {
-                return globalJobListeners[name];
+                throw new ArgumentNullException(nameof(name));
+            }
+
+            lock (globalJobListenerLock)
+            {
+                // Avoid initializing globalJobListeners when no job listeners have been added
+                if (globalJobListeners == null || !globalJobListeners.TryGetValue(name, out var jobListener))
+                {
+                    throw new KeyNotFoundException();
+                }
+
+                return jobListener;
             }
         }
 
@@ -189,7 +297,7 @@ namespace Quartz.Core
             lock (globalTriggerListeners)
             {
                 globalTriggerListeners[triggerListener.Name] = triggerListener;
-                var matchers = new List<IMatcher<TriggerKey>> {matcher};
+                var matchers = new List<IMatcher<TriggerKey>> { matcher };
                 globalTriggerListenersMatchers[triggerListener.Name] = matchers;
             }
         }
@@ -305,9 +413,25 @@ namespace Quartz.Core
         {
             lock (schedulerListeners)
             {
-                return schedulerListeners.Count > 0 
+                return schedulerListeners.Count > 0
                     ? new List<ISchedulerListener>(schedulerListeners)
                     : EmptyReadOnlyCollection<ISchedulerListener>.Instance;
+            }
+        }
+
+        private void RemoveJobListenerMatchers(string listenerName)
+        {
+            if (globalJobListenersMatchers == null)
+            {
+                return;
+            }
+
+            // If we're removing the last matcher of the only job listener with matchers, then
+            // reset globalJobListenersMatchers to null to avoid having to lock in subsequent calls
+            // to GetJobListenerMatchers(string listenerName)
+            if (globalJobListenersMatchers.Remove(listenerName) && globalJobListenersMatchers.Count == 0)
+            {
+                globalJobListenersMatchers = null;
             }
         }
     }