… src/Quartz.Benchmark/SchedulerBenchmark.cs
… src/Quartz.Tests.Unit/Core/ListenerManagerTest .cs
… src/Quartz/Core/QuartzScheduler.cs
… src/Quartz/IListenerManager.cs