This is a draft PR that mostly aims to gather feedback on the changes that I applied as I'm sure some (most? :p) of these changes will probably be considered controversial.

There are two parts to this PR:
* Optimize the job listener part of ListenerManagerImpl.
* Optimize the job listener part of QuartzScheduler.

**ListenerManagerImpl**
Optimize the way we deal with job listeners by:
* eliminating locking in GetJobListenerMatchers when no matchers were were added.
* eliminating locking in GetJobListeners when no job listeners were added.
* lazily initializing globalJobListeners and globalJobListenersMatchers dictionaries.
* not adding a EverythingMatcher for a job listener when no matchers were specified.

<details>
<summary>Benchmark results</summary>

|                                                        Method | Branch |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|-------------------------------------------------------------- |--------|---------:|---------:|---------:|-------:|----------:|
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | v4     | 402.6 ns |  3.26 ns |  2.72 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | pr     | 350.2 ns |  0.67 ns |  0.63 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | v4     | 467.0 ns |  6.56 ns |  5.81 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | pr     | 473.5 ns |  4.97 ns |  4.65 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | v4     | 398.1 ns |  0.69 ns |  0.64 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | pr     | 348.3 ns |  1.14 ns |  1.01 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | v4     | 472.5 ns |  9.17 ns |  9.42 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | pr     | 474.7 ns |  3.21 ns |  3.00 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | v4     | 792.0 ns |  3.59 ns |  3.18 ns | 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | pr     | 713.3 ns | 14.24 ns | 13.32 ns | 0.0896 |     377 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | v4     | 765.7 ns | 15.00 ns | 20.53 ns | 0.1075 |     451 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | pr     | 611.3 ns |  9.01 ns | 12.63 ns | 0.0900 |     379 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | v4     | 789.2 ns |  2.64 ns |  2.47 ns | 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | pr     | 702.4 ns |  5.84 ns |  4.56 ns | 0.0896 |     377 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | v4     | 737.1 ns | 14.70 ns | 20.61 ns | 0.1075 |     452 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | pr     | 581.9 ns | 11.61 ns | 20.64 ns | 0.0900 |     379 B |
</details>

_Conclusion:_
This part of the PR leads to nice improvements, but nothing too spectacular.

**QuartzScheduler**
Optimize notify of job listeners by:
* only special casing ExecutingJobsManager as "internal" job listener.
* no longer checking for matchers in case of ExecutingJobsManager.
* removing methods to add, remove or get internal job listeners.
* no longer add the job store as internal listener if it implements IJobListener.

I also removed the protected ctor of **QuartzScheduler**, but that can easily be reverted.

<details>
<summary>Benchmark results</summary>

|                                                        Method | Branch |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|-------------------------------------------------------------- |--------|---------:|--------:|--------:|-------:|----------:|
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | v4     | 402.6 ns | 3.26 ns |  2.72 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | pr     | 105.3 ns | 0.26 ns | 0.23 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | v4     | 467.0 ns | 6.56 ns |  5.81 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | pr     | 143.5 ns | 1.27 ns | 1.19 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | v4     | 398.1 ns | 0.69 ns |  0.64 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | pr     | 130.9 ns | 0.70 ns | 0.59 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | v4     | 472.5 ns | 9.17 ns |  9.42 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | pr     | 142.8 ns | 1.20 ns | 1.12 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | v4     | 792.0 ns | 3.59 ns |  3.18 ns| 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | pr     | 372.7 ns | 0.99 ns | 0.77 ns | 0.0381 |     160 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | v4     | 765.7 ns |15.00 ns | 20.53 ns| 0.1075 |     451 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | pr     | 328.1 ns | 3.46 ns | 3.06 ns | 0.0381 |     161 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | v4     | 789.2 ns | 2.64 ns |  2.47 ns| 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | pr     | 371.5 ns | 1.10 ns | 0.92 ns | 0.0381 |     160 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | v4     | 737.1 ns |14.70 ns | 20.61 ns| 0.1075 |     452 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | pr     | 322.1 ns | 3.31 ns | 2.76 ns | 0.0381 |     161 B |
</details>

_Conclusion:_
If we combine the changes to **ListenerManagerImpl** and **QuartzScheduler**, we get great improvements on both performance and allocations.