@drieseng 2022-04-20
Personally, I think it's ok to remove the methods to add/remove/get the internal listeners.
The "internal" part of the name says enough.

No longer automatically (or should I say "automagically") registering the job store as a job listener is also a positive (although breaking) change in my book. But you may be reading another book of course :-)