@drieseng 2022-04-22
@lahma I was able to improve performance and reduce allocations even further by changing `IListenerManager.GetJobListeners()` to return **IJobListener[]** instead of **IReadOnlyCollection<IJobListener>**:

|                                                        Method | Branch |     Mean |   Error |  StdDev |  Gen 0 | Allocated |
|-------------------------------------------------------------- |--------|---------:|--------:|--------:|-------:|----------:|
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | v4     | 402.6 ns | 3.26 ns |  2.72 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | pr     | 104.5 ns | 0.48 ns | 0.45 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | v4     | 467.0 ns | 6.56 ns |  5.81 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | pr     | 139.5 ns | 1.53 ns | 1.43 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | v4     | 398.1 ns | 0.69 ns |  0.64 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | pr     | 101.7 ns | 0.24 ns | 0.23 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | v4     | 472.5 ns | 9.17 ns |  9.42 ns| 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | pr     | 142.1 ns | 1.65 ns | 1.46 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | v4     | 792.0 ns | 3.59 ns |  3.18 ns| 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | pr     | 339.3 ns | 1.01 ns | 0.94 ns | 0.0191 |      80 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | v4     | 765.7 ns |15.00 ns | 20.53 ns| 0.1075 |     451 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | pr     | 279.2 ns | 4.30 ns | 4.02 ns | 0.0190 |      81 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | v4     | 789.2 ns | 2.64 ns |  2.47 ns| 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | pr     | 339.4 ns | 0.80 ns | 0.75 ns | 0.0191 |      80 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | v4     | 737.1 ns |14.70 ns | 20.61 ns| 0.1075 |     452 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | pr     | 284.2 ns | 4.29 ns | 4.02 ns | 0.0189 |      81 B |