@drieseng 2022-04-27
@lahma I'm not avoid creation of lists (and avoiding locking) for the case where there are lots of schedulers; I avoid paying the cost for supporting "custom" job listeners when the user did not register any such listeners. At the same time, I also managed to reduce the cost of actually having custom job listeners.

I'll run more benchmarks later this week. When all are satisfactory and I updated the changelog, I'll remove the "draft" status from this PR.