@drieseng 2022-05-02
@lahma, here are the final benchmark results:

**JobRunShellBenchmark**

|                                                      Method | Branch |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
|------------------------------------------------------------ |-------:|-----------|----------:|----------:|-------:|----------:|
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | pr     |  1.322 us | 0.0262 us | 0.0245 us | 0.1163 |     489 B |
| Success_NoTriggerListenersAndSingleJobListener_MayFireAgain | v4     |  1.894 us | 0.0083 us | 0.0074 us | 0.1850 |     778 B |

**QuartSchedulerBenchmark**

|                                                        Method | Branch |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|-------------------------------------------------------------- |-------:|---------:|---------:|---------:|-------:|----------:|
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | pr     | 103.0 ns |  0.17 ns |  0.15 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartScheduler1_SingleThreaded | v4     | 396.7 ns |  0.91 ns |  0.76 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | pr     | 136.1 ns |  1.36 ns |  1.27 ns | 0.0114 |      48 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler1_MultiThreaded | v4     | 477.5 ns |  5.44 ns |  5.09 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | pr     | 352.7 ns |  0.57 ns |  0.51 ns | 0.0267 |     112 B |
| NotifyJobListenersToBeExecuted_QuartScheduler2_SingleThreaded | v4     | 792.6 ns |  2.46 ns |  2.18 ns | 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | pr     | 317.8 ns |  2.10 ns |  1.96 ns | 0.0269 |     113 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler2_MultiThreaded | v4     | 719.3 ns | 14.29 ns | 20.50 ns | 0.1075 |     451 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | pr     | 410.4 ns |  0.88 ns |  0.82 ns | 0.0439 |     185 B |
| NotifyJobListenersToBeExecuted_QuartScheduler3_SingleThreaded | v4     | 850.0 ns |  3.18 ns |  2.65 ns | 0.1068 |     449 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | pr     | 413.9 ns |  6.38 ns |  5.33 ns | 0.0442 |     185 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler3_MultiThreaded | v4     | 765.0 ns | 15.25 ns | 21.38 ns | 0.1075 |     451 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | pr     | 533.2 ns |  0.30 ns |  0.25 ns | 0.0458 |     193 B |
| NotifyJobListenersToBeExecuted_QuartScheduler4_SingleThreaded | v4     | 989.9 ns | 19.46 ns | 18.20 ns | 0.1259 |     530 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | pr     | 564.1 ns |  7.49 ns |  7.00 ns | 0.0460 |     194 B |
| NotifyJobListenersToBeExecuted_QuartzScheduler4_MultiThreaded | v4     | 912.8 ns | 17.99 ns | 34.66 ns | 0.1267 |     532 B |

I've marked this PR a ready for review now.