@drieseng 2022-05-02
@lahma The (async) tests that I added are failing on GitHub (while passing locally). I had a similar issue before for which I also registered an issue with **NUnit**, but they couldn't offer me a solution. Let me rewrite those tests to have non-async test methods :(