@drieseng 2022-05-03
@lahma Retargeting the PR using the GitHub functionality didn't work out well. Finally managed to get it done though.

I plan to follow up on this PR with similar changes for scheduler and trigger listeners (if proven to have a positive effect on performance and allocations).

When are you planning to release a beta for 4.0?

PS. I suck at git :p