@lahma 2022-05-03
> @lahma Retargeting the PR using the GitHub functionality didn't work out well. Finally managed to get it done though.

Thanks for doing the dance and finishing this quality PR!

> I plan to follow up on this PR with similar changes for scheduler and trigger listeners (if proven to have a positive effect on performance and allocations).

Sounds good.

> When are you planning to release a beta for 4.0?

All main builds should be available on MyGet so I'm not in a hurry. There's a lot of API breaking to do...

> PS. I suck at git :p

Git can be a beast... 😉